+++ example/lib/app/modules/cart/controller.dart
@@ -1,6 +1,5 @@
 import 'package:example/app/data/models/item.dart';
 import 'package:minimals_state_manager/app/state/min_notifier.dart';
-import 'package:minimals_state_manager/app/observers/min_app_lifecycle.dart';
 
 class CartController extends MinNotifier {
   @override
@@ -15,23 +14,18 @@ class CartController extends MinNotifier {
 
   List<Item> items = <Item>[];
 
-  bool addItem(Item item) {
+  void addItem(Item item) {
     bool exists = items.any((i) => i.id == item.id);
 
     if (exists) {
-      return true;
     } else {
       items.add(item);
-      items = List.of(items);
       notifyListeners();
-      return true;
     }
   }
 
-  bool removeItem(item) {
+  void removeItem(item) {
     items.removeWhere((element) => element == item);
-    items = List.of(items);
     notifyListeners();
-    return true;
   }
 }