+++ example/lib/app/modules/home/widgets/drawer.dart
@@ -1,19 +1,24 @@
+import 'package:example/app/data/services/auth/service.dart';
+import 'package:example/app/data/services/setup_locator.dart';
 import 'package:flutter/material.dart';
 
 class HomeDrawer extends StatelessWidget {
-  const HomeDrawer({super.key});
-
+  HomeDrawer({super.key});
+  final AuthService _authService = min.get<AuthService>();
   @override
   Widget build(BuildContext context) {
     return Drawer(
       child: ListView(
         padding: EdgeInsets.zero,
         children: <Widget>[
-          const DrawerHeader(
-            decoration: BoxDecoration(
+          DrawerHeader(
+            decoration: const BoxDecoration(
               color: Colors.blue,
             ),
-            child: Text('Drawer Header'),
+            child: Text(
+              _authService.user.name!,
+              style: const TextStyle(color: Colors.white),
+            ),
           ),
           ListTile(
             title: const Text('Item 1'),