+++ example/lib/app/modules/home/widgets/filters.dart
@@ -1,19 +1,17 @@
+import 'package:example/app/data/enums/item_type.dart';
 import 'package:example/app/modules/home/controller.dart';
 import 'package:flutter/material.dart';
-import 'package:minimals_state_manager/app/extensions/min_provider_extensions.dart';
 import 'package:minimals_state_manager/app/widgets/min_selector.dart';
 
 class BSFilters extends StatelessWidget {
-  const BSFilters({super.key});
+  const BSFilters({super.key, required this.controller});
+  final HomeController controller;
 
   @override
   Widget build(BuildContext context) {
-    final controller = context.read<HomeController>();
-
     return Container(
       padding: const EdgeInsets.all(16.0),
       decoration: const BoxDecoration(
-        // color: Colors.grey,
         borderRadius: BorderRadius.all(
           Radius.circular(6.0),
         ),
@@ -35,28 +33,29 @@ class BSFilters extends StatelessWidget {
             flex: 4,
             child: SizedBox(
               width: MediaQuery.of(context).size.width * 0.5,
-              child: $<HomeController, int>(
+              child: $<HomeController, ItemType>(
                 notifier: controller,
                 selector: (notifier) => notifier.filter,
                 builder: (context, filter) => ListView.builder(
-                  // scrollDirection: Axis.horizontal,
-                  itemCount: 3,
+                  itemCount: ItemType.validTypes.length,
                   itemBuilder: (context, index) => InkWell(
-                    onTap: () => controller.changeFilter(index + 1),
+                    onTap: () =>
+                        controller.changeFilter(ItemType.validTypes[index]),
                     child: Container(
                       margin: const EdgeInsets.symmetric(vertical: 4.0),
                       height: MediaQuery.of(context).size.height * 0.06,
                       decoration: BoxDecoration(
                         border: Border.all(
-                            color:
-                                controller.filter != 0 && (index + 1 == filter)
-                                    ? Colors.green
-                                    : Colors.amber),
+                            color: filter != ItemType.EMPTY &&
+                                    (ItemType.validTypes[index] == filter)
+                                ? Colors.green
+                                : Colors.amber),
                         borderRadius: const BorderRadius.all(
                           Radius.circular(6.0),
                         ),
                       ),
-                      child: Center(child: Text('Type : ${index + 1}')),
+                      child:
+                          Center(child: Text(ItemType.validTypes[index].name)),
                     ),
                   ),
                 ),
@@ -65,23 +64,28 @@ class BSFilters extends StatelessWidget {
           ),
           Flexible(
             child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceAround,
+              spacing: 32.0,
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
               children: [
-                MaterialButton(
-                  minWidth: MediaQuery.of(context).size.width * 0.3,
-                  color: Colors.redAccent,
-                  onPressed: () => Navigator.pop(context),
-                  child: const Text('CLOSE'),
+                Expanded(
+                  child: MaterialButton(
+                    minWidth: MediaQuery.of(context).size.width * 0.3,
+                    color: Colors.redAccent,
+                    onPressed: () => Navigator.pop(context),
+                    child: const Text('CLOSE'),
+                  ),
                 ),
-                MaterialButton(
-                  padding: EdgeInsets.zero,
-                  minWidth: MediaQuery.of(context).size.width * 0.3,
-                  color: Colors.amberAccent,
-                  onPressed: () async {
-                    await controller.removeFilters();
-                    context.mounted ? Navigator.pop(context) : null;
-                  },
-                  child: const Text('REMOVE FILTERS'),
+                Expanded(
+                  child: MaterialButton(
+                    padding: EdgeInsets.zero,
+                    minWidth: MediaQuery.of(context).size.width * 0.3,
+                    color: Colors.amberAccent,
+                    onPressed: () async {
+                      await controller.removeFilters();
+                      context.mounted ? Navigator.pop(context) : null;
+                    },
+                    child: const Text('REMOVE FILTERS'),
+                  ),
                 ),
               ],
             ),