+++ example/lib/app/modules/home/widgets/item.dart
@@ -1,7 +1,6 @@
 import 'package:example/app/data/models/item.dart';
 import 'package:example/app/modules/cart/controller.dart';
 import 'package:example/app/modules/home/controller.dart';
-import 'package:example/app/modules/widgets/build_banner.dart';
 import 'package:flutter/material.dart';
 import 'package:minimals_state_manager/app/extensions/min_provider_extensions.dart';
 import 'package:minimals_state_manager/app/provider/min_multi_provider.dart';
@@ -22,9 +21,12 @@ class ItemWidget extends Container {
         margin: const EdgeInsets.all(6.0),
         height: 100.0,
         width: MediaQuery.of(context).size.width,
-        decoration: BoxDecoration(
-            color: Colors.grey.withOpacity(.5),
-            borderRadius: const BorderRadius.all(Radius.circular(4.0))),
+        decoration: const BoxDecoration(
+          color: Colors.white,
+          borderRadius: BorderRadius.all(
+            Radius.circular(4.0),
+          ),
+        ),
         child: Row(
           mainAxisAlignment: MainAxisAlignment.spaceBetween,
           children: [
@@ -33,18 +35,13 @@ class ItemWidget extends Container {
               mainAxisAlignment: MainAxisAlignment.spaceAround,
               children: [
                 Text('${items[index].name} - $index'),
-                Text('Type ${items[index].type}'),
-                Text('${items[index].value}'),
+                Text(items[index].type.name),
+                Text('R\$ ${items[index].value}'),
               ],
             ),
             IconButton(
               onPressed: () {
-                if (cartController.addItem(items[index])) {
-                  ScaffoldMessenger.of(context).showMaterialBanner(buildBanner(
-                      context,
-                      message: 'Add item to cart',
-                      isRemoved: false));
-                }
+                cartController.addItem(items[index]);
               },
               icon: const Icon(
                 Icons.add,